renaming packages and interfaces in new COM bits
