+++ contrib/platform/src/com/sun/jna/platform/win32/COM/util/annotation/ComMethod.java
@@ -10,7 +10,7 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  * Lesser General Public License for more details.
  */
-package com.sun.jna.platform.win32.COM.proxy.annotation;
+package com.sun.jna.platform.win32.COM.util.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Inherited;
@@ -19,9 +19,9 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target({ ElementType.TYPE })
+@Target({ ElementType.METHOD })
 @Inherited
-public @interface ComObject {
-	String clsId() default "";
-	String progId() default "";
+public @interface ComMethod {
+	String iid() default "";
+	String name() default "";
 }
