+++ contrib/platform/src/com/sun/jna/platform/win32/COM/util/annotation/ComProperty.java
@@ -10,7 +10,7 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  * Lesser General Public License for more details.
  */
-package com.sun.jna.platform.win32.COM.proxy.annotation;
+package com.sun.jna.platform.win32.COM.util.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Inherited;
@@ -21,7 +21,7 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ ElementType.METHOD })
 @Inherited
-public @interface Property {
+public @interface ComProperty {
 	String iid() default "";
 	String name() default "";
 }
