+++ contrib/platform/test/com/sun/jna/platform/win32/COM/EnumMoniker_Test.java
@@ -26,7 +26,7 @@ import com.sun.jna.platform.win32.WinDef.DWORD;
 import com.sun.jna.platform.win32.WinDef.ULONG;
 import com.sun.jna.platform.win32.WinDef.ULONGByReference;
 import com.sun.jna.platform.win32.WinNT.HRESULT;
-import com.sun.jna.platform.win32.COM.proxy.ProxyObject;
+import com.sun.jna.platform.win32.COM.util.ProxyObject;
 import com.sun.jna.ptr.PointerByReference;
 
 public class EnumMoniker_Test {
