+++ eclipse/com.sun.jna.core/.settings/org.eclipse.pde.core.prefs
@@ -1,0 +1,2 @@
+BUNDLE_ROOT_PATH=src/main/resources
+eclipse.preferences.version=1
