+++ eclipse/com.sun.jna.feature/build.properties
@@ -1,2 +1,2 @@
 bin.includes = feature.xml,\
-               LICENSE.txt
+               licence.txt
