+++ eclipse/com.sun.jna.platform/pom.xml
@@ -16,11 +16,20 @@
 	</properties>
 	
 	<dependencies>
+		
 		<dependency>
 			<groupId>com.sun.jna</groupId>
 			<artifactId>com.sun.jna.core</artifactId>
 			<version>${project.version}</version>		
 		</dependency>
+		
+		<dependency>
+    		<groupId>junit</groupId>
+    		<artifactId>junit</artifactId>
+    		<version>4.11</version>
+    		<scope>test</scope>
+		</dependency>
+
 	</dependencies>
